Matching "jaudio/jammain_2.c" for more revisions

+++ include/jaudio/jammain_2.h
@@ -305,7 +305,7 @@ void Jam_CheckExportApp(void);
 void Jam_CheckImportApp(void);
 void Jam_WritePortIndirect(void);
 void Jam_ReadPortIndirect(void);
-BOOL Jam_CheckPortIndirect(seqp_*, u8, int);
+BOOL Jam_CheckPortIndirect(seqp_*, u32, u16);
 BOOL Jam_WritePortAppDirect(seqp_*, u32, u16);
 BOOL Jam_ReadPortAppDirect(seqp_*, u32, u16*);
 BOOL Jam_CheckPortAppDirect(seqp_*, u32, u16);
